requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
gspread==5.7.2
google-auth==2.17.3
tqdm==4.64.1
python-telegram-bot==21.5
ratelimit==2.2.1